backend/pyproject.toml
[project]
name = "readmatrix"
version = "0.1.0"
description = "Local-first personal knowledge platform with grounded Q&A"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-frontmatter>=1.1.0",
    "chromadb==1.4.1",
    "openai>=1.10.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.8.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[project.scripts]
readmatrix = "readmatrix.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
